refact: monitor api (#11)
